requirements.txt
#Libraries needed for this project as of 13/11/2020
#pip freeze to see any missing libaries and then install any that you are missing 

asgiref==3.3.1
certifi==2020.6.20
Django==3.1.3
django-crispy-forms==1.9.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
wincertstore==0.2